feat(user): fractional GB limits + lifetime traffic + two-pane edit dialog
Traffic limit and emergency per-window quota now accept decimal GB (>=0,
e.g. 0.3). Backend TrafficLimitGB / EmergencyAccessQuotaGB were integers
and rejected fractional JSON; switched to float64 with int64() byte casts,
floatField KV serialization, and step="any" inputs with a >=0-only check.

User management gains lifetime traffic detail. The edit dialog is reworked
into a Cloudreve-style two-pane layout: left aggregates identity (avatar,
role/SSO/status badges, ID/UUID, a usage bar with period-used / limit /
lifetime totals plus up/down breakdown, created-at, copy-sub); right is the
editable form grid. User DTO adds read-only lifetime_{up,down,total}_bytes.

+++ internal/ports/repos.go
@@ -370,7 +370,8 @@ type UISettings struct {
 	// can consume on top of the user's already-exceeded period. 0 = unlimited
 	// (only the time/count limits apply). When the user crosses the quota, the
 	// traffic poll ends the emergency window early and re-runs auto-disable.
-	EmergencyAccessQuotaGB int `json:"emergency_access_quota_gb"`
+	// Fractional GB allowed (e.g. 0.12). Stored as a float string in the KV.
+	EmergencyAccessQuotaGB float64 `json:"emergency_access_quota_gb"`
 
 	// ---- Subscription settings ----
 	// SubPath is the URL path prefix for subscription endpoints.
